pyproject.toml
[project]
name = "tom-camp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.17"
dependencies = [
    "coverage>=7.8.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
